requirements.txt
aiofiles>=23.2.1
aiogram>=3.11.0
openpyxl>=3.1.5
python-dotenv>=1.0.1
requests>=2.32.3